requirements.txt
pandas
scikit-learn
joblib
fastapi
uvicorn
pydantic
python-jose[cryptography]
bcrypt
SQLAlchemy
pydantic-settings
onnxruntime
transformers
torch
